% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquisition_functions.R
\name{acq_GP_MI}
\alias{acq_GP_MI}
\title{Acquisition Function GP-MI (Mutual Information)}
\usage{
acq_GP_MI(delta = 1e-06, alpha = log(2/delta), kappa = sqrt(alpha))
}
\description{
Acquisition Function GP-MI (Mutual Information)
}

