% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquisition_functions.R
\name{acq_GP_UCB}
\alias{acq_GP_UCB}
\title{Acquisition Function GP-UCB (Upper Confidence Bound)}
\usage{
acq_GP_UCB(beta = log(2/1e-06), kappa = sqrt(beta))
}
\arguments{
\item{beta}{a scalar.}

\item{kappa}{a scalar.}
}
\value{
acquisition function GP-UCB
}
\description{
Acquisition Function GP-UCB (Upper Confidence Bound)
}

