% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesopt2.R
\name{bayesopt2}
\alias{bayesopt2}
\title{Bayesian Optimization}
\usage{
bayesopt2(objective_func, ..., iter = 10, kernel = c("squared_exponential"),
  acq_func = acq_GP_UCB())
}
\arguments{
\item{objective_func}{a function.}

\item{...}{list of parameters.}

\item{iter}{an integer.}

\item{kernel}{a character string.}

\item{acq_func}{a function.}
}
\value{
a list
}
\description{
Bayesian Optimization
}

