% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesopt3.R
\name{bayesopt3}
\alias{bayesopt3}
\title{Bayesian Optimization}
\usage{
bayesopt3(objective_func, ..., iter = 10)
}
\arguments{
\item{objective_func}{a function.}

\item{...}{list of parameters.}

\item{iter}{an integer.}

\item{noise}{logical.}

\item{kernel}{a character string.}

\item{acq_func}{a function.}
}
\value{
a list
}
\description{
Bayesian Optimization
}

