% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_matrix.R
\name{compute_covariance_matrix}
\alias{compute_covariance_matrix}
\title{Compute covaiance matrix}
\usage{
compute_covariance_matrix(x_mat, nu, theta, theta0,
  kernel_func = kernel_gp_squared_exponential)
}
\arguments{
\item{x_mat}{a matrix.}

\item{nu}{a scalar.}

\item{theta}{a numeric vector.}

\item{theta0}{a numeric vector.}

\item{kernel_func}{a function.}
}
\value{
covariance matrix
}
\description{
Compute covaiance matrix
}

