% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_hyparam.R
\name{estimate_hyperparam}
\alias{estimate_hyperparam}
\title{Estimate Hyper Parameters of GP by Empirical Bayes Method}
\usage{
estimate_hyperparam(y, x_mat, kernel_func = kernel_gp_squared_exponential)
}
\description{
Estimate Hyper Parameters of GP by Empirical Bayes Method
}

