% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{Himmelblau}
\alias{Himmelblau}
\title{Himmelblau's function}
\usage{
Himmelblau(x, y)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}
}
\value{
value
}
\description{
Himmelblau's function
}

