% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_gp_functions.R
\name{kernel_gp_squared_exponential}
\alias{kernel_gp_squared_exponential}
\title{Squared exponential kernel for GP}
\usage{
kernel_gp_squared_exponential(x1, x2, theta)
}
\arguments{
\item{x1}{a numeric vector.}

\item{x2}{a numeric vector.}

\item{theata}{a numeric vetor.}
}
\description{
Squared exponential kernel for GP
}

