% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesopt.R
\name{bayesopt}
\alias{bayesopt}
\title{Bayesian Optimization}
\usage{
bayesopt(objective_func, ..., iter = 10, kernel = c("matern5_2",
  "matern3_2", "square_exp"), acq_func = c("mutual_information",
  "confidence_bound"), k = sqrt(log(2/(1e-06))), plot = FALSE)
}
\arguments{
\item{objective_func}{a function.}

\item{...}{list of parameters.}

\item{iter}{an integer.}

\item{kernel}{a character string.}

\item{acq_func}{a character string.}

\item{k}{a number.}

\item{plot}{logical.}
}
\value{
a list
}
\description{
Bayesian Optimization
}

