% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_matrix.R
\name{compute_covariance_matrix}
\alias{compute_covariance_matrix}
\title{Compute covaiance matrix}
\usage{
compute_covariance_matrix(x_mat, theta,
  kernel_func = kernel_gp_squared_exponential)
}
\value{
covariance matrix
}
\description{
Compute covaiance matrix
}

