% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_hyparam.R
\name{estimate_hyperparam}
\alias{estimate_hyperparam}
\title{Estimate Hyper Parameters of GP by Empirical Bayes Method}
\usage{
estimate_hyperparam(x_mat, y, kernel_func = kernel_gp_squared_exponential)
}
\arguments{
\item{x_mat}{a matrix.}

\item{y}{a numeric vector.}

\item{kernel_func}{a function.}
}
\description{
Estimate Hyper Parameters of GP by Empirical Bayes Method
}

