% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_gp_functions.R
\name{kernel_gp_exponential}
\alias{kernel_gp_exponential}
\title{Exponential kernel for GP}
\usage{
kernel_gp_exponential(x1, x2, theta, power = 2)
}
\arguments{
\item{x1}{a numeric vector.}

\item{x2}{a numeric vector.}

\item{theta}{a numeric vetor.}

\item{power}{a scalar.}
}
\description{
Exponential kernel for GP
}

